更严格的类型转换限制

+++ src/component/comp_cf.cpp
@@ -11,7 +11,7 @@ SecOrderFunction::SecOrderFunction(SecOrderFunction::Param& param)
 
 float SecOrderFunction::GetValue(float fb) {
   float out =
-      this->param_.a * pow(fb, 2) + this->param_.b * fb + this->param_.c;
+      this->param_.a * powf(fb, 2) + this->param_.b * fb + this->param_.c;
   clampf(&out, this->param_.min, this->param_.max);
   return out;
 }
